address comments; restructure op files; more tests.
